refactor(auth): trigger webhook token migration via @Subscriber('rundeck.bootstrap')
